test: regression pins, folder consolidation, parity residuals
- Pin the quality gate: app/navigate file-traversal + malformed-sha rejection
  (adeRpcServer), ade link round-trip refusal (CLI), lane search envelope
  (searchService).
- Consolidate services/deeplinks to one test file (dispatch contract merged
  into protocolHandler.test.ts).
- Parity residuals: search README deepLink prose, ios-companion universal-link
  notes, ade-cli README CLI-surface inventory.
